CIV-6299 implement strategy pattern for the handler
