+++ src/main/java/uk/gov/hmcts/reform/civil/enums/CaseState.java
@@ -14,6 +14,7 @@ public enum CaseState {
     PREPARE_FOR_HEARING_CONDUCT_HEARING,
     DECISION_OUTCOME,
     IN_MEDIATION,
-    CASE_SETTLED
+    CASE_SETTLED,
+    CASE_STAYED
 }
