+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/strategy/UploadTranslatedDocumentV1Strategy.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.reform.civil.handler.callback.user.strategy;
+
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.civil.callback.CallbackParams;
+
+@Component
+public class UploadTranslatedDocumentV1Strategy implements UploadTranslatedDocumentStrategy{
+    @Override
+    public CallbackResponse uploadDocument(CallbackParams callbackParams) {
+        return null;
+    }
+}
